+++ pct_scalewiz/components/project_info.py
@@ -19,86 +19,86 @@ class ProjectInfo(ttk.Frame):
 
         # row 0 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Customer:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.customer)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.customer)
         parent.render(lbl, ent, 0)
 
         # row 1 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Submitted by:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.submitted_by)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.submitted_by)
         parent.render(lbl, ent, 1)
 
         # row 2 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Production company:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.client)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.client)
         parent.render(lbl, ent, 2)
 
         # row 3 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Field:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.field)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.field)
         parent.render(lbl, ent, 3)
 
         # row 4 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Sample point:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.sample)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.sample)
         parent.render(lbl, ent, 4)
 
         # row 5 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Project name:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.name)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.name)
         parent.render(lbl, ent, 5)
 
         # row 6 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Date sampled:")
         # this is to refresh the entry later
-        current_value = parent.editorProject.sample_date.get()
+        current_value = parent.editor_project.sample_date.get()
         ent = tkcal.DateEntry(
-            self, textvariable=parent.editorProject.sample_date, date_pattern="mm/dd/Y"
+            self, textvariable=parent.editor_project.sample_date, date_pattern="mm/dd/Y"
         )
-        lbl.bind("<Button-1>", lambda _: parent.editorProject.sample_date.set(""))
-        parent.editorProject.sample_date.set(current_value)
+        lbl.bind("<Button-1>", lambda _: parent.editor_project.sample_date.set(""))
+        parent.editor_project.sample_date.set(current_value)
         parent.render(lbl, ent, 6)
 
         # row 7 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Date received:")
-        current_value = parent.editorProject.received_date.get()
+        current_value = parent.editor_project.received_date.get()
         ent = tkcal.DateEntry(
             self,
-            textvariable=parent.editorProject.received_date,
+            textvariable=parent.editor_project.received_date,
             date_pattern="mm/dd/yyyy",
         )
-        lbl.bind("<Button-1>", lambda _: parent.editorProject.received_date.set(""))
-        parent.editorProject.received_date.set(current_value)
+        lbl.bind("<Button-1>", lambda _: parent.editor_project.received_date.set(""))
+        parent.editor_project.received_date.set(current_value)
         parent.render(lbl, ent, 7)
 
         # row 8 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Date completed:")
-        current_value = parent.editorProject.completed_date.get()
+        current_value = parent.editor_project.completed_date.get()
         ent = tkcal.DateEntry(
             self,
-            textvariable=parent.editorProject.completed_date,
+            textvariable=parent.editor_project.completed_date,
             date_pattern="mm/dd/yyyy",
         )
-        lbl.bind("<Button-1>", lambda _: parent.editorProject.completed_date.set(""))
-        parent.editorProject.completed_date.set(current_value)
+        lbl.bind("<Button-1>", lambda _: parent.editor_project.completed_date.set(""))
+        parent.editor_project.completed_date.set(current_value)
         parent.render(lbl, ent, 8)
 
         # row 9 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Analyst:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.analyst)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.analyst)
         parent.render(lbl, ent, 9)
 
         # row 10 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Analysis number(s):")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.numbers)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.numbers)
         parent.render(lbl, ent, 10)
 
         # row 11 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Notes:")
-        ent = ttk.Entry(self, textvariable=parent.editorProject.notes)
+        ent = ttk.Entry(self, textvariable=parent.editor_project.notes)
         parent.render(lbl, ent, 11)
 
         # not implemented
         # row 12 ---------------------------------------------------------------
         # pathLbl = ttk.Label(self, text="File path:")
-        # pathEnt = ttk.Entry(self, textvariable=parent.editorProject.path)
+        # pathEnt = ttk.Entry(self, textvariable=parent.editor_project.path)
         # parent.render(pathLbl, pathEnt, 12)