+++ pct_scalewiz/components/project_params.py
@@ -16,7 +16,7 @@ class ProjectParams(ttk.Frame):
         # row 0 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Bicarbonates (mg/L):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.bicarbs, from_=0, to=999999
+            self, textvariable=parent.editor_project.bicarbs, from_=0, to=999999
         )
         parent.render(lbl, ent, 0)
 
@@ -26,58 +26,64 @@ class ProjectParams(ttk.Frame):
         ent.grid_columnconfigure(0, weight=1)
         ent.grid_columnconfigure(1, weight=1)
         ttk.Radiobutton(
-            ent, text="Yes", variable=parent.editorProject.bicarbs_increased, value=True
+            ent,
+            text="Yes",
+            variable=parent.editor_project.bicarbs_increased,
+            value=True,
         ).grid(row=0, column=0)
         ttk.Radiobutton(
-            ent, text="No", variable=parent.editorProject.bicarbs_increased, value=False
+            ent,
+            text="No",
+            variable=parent.editor_project.bicarbs_increased,
+            value=False,
         ).grid(row=0, column=1)
         parent.render(lbl, ent, 1)
 
         # row 2 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Chlorides (mg/L):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.chlorides, from_=0, to=999999
+            self, textvariable=parent.editor_project.chlorides, from_=0, to=999999
         )
         parent.render(lbl, ent, 2)
 
         # row 3 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Test temperature (°F):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.temperature, from_=0, to=9999
+            self, textvariable=parent.editor_project.temperature, from_=0, to=9999
         )
         parent.render(lbl, ent, 3)
 
         # row 4 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Baseline pressure (PSI):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.baseline, from_=0, to=9999
+            self, textvariable=parent.editor_project.baseline, from_=0, to=9999
         )
         parent.render(lbl, ent, 4)
 
         # row 5 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Limiting pressure (PSI):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.limit_psi, from_=0, to=9999
+            self, textvariable=parent.editor_project.limit_psi, from_=0, to=9999
         )
         parent.render(lbl, ent, 5)
 
         # row 6 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Time limit (min.):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.limit_minutes, from_=0, to=9999
+            self, textvariable=parent.editor_project.limit_minutes, from_=0, to=9999
         )
         parent.render(lbl, ent, 6)
 
         # row 7 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Reading interval (s):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.interval, from_=0, to=9999
+            self, textvariable=parent.editor_project.interval, from_=0, to=9999
         )
         parent.render(lbl, ent, 7)
 
         # row 8 ---------------------------------------------------------------
         lbl = ttk.Label(self, text="Uptake time (s):")
         ent = ttk.Spinbox(
-            self, textvariable=parent.editorProject.uptake, from_=0, to=9999
+            self, textvariable=parent.editor_project.uptake, from_=0, to=9999
         )
         parent.render(lbl, ent, 8)