+++ pct_scalewiz/models/project.py
@@ -6,9 +6,7 @@ import json
 import logging
 import os
 import tkinter as tk
-import typing
 
-from pct_scalewiz.helpers.get_resource import get_resource
 from pct_scalewiz.helpers.sort_nicely import sort_nicely
 from pct_scalewiz.models.test import Test
 
@@ -53,10 +51,10 @@ class Project:
         self.tests: list[dict] = []
 
         # maintain live fields
-        self.customer.trace("w", self.make_name)
-        self.client.trace("w", self.make_name)
-        self.field.trace("w", self.make_name)
-        self.sample.trace("w", self.make_name)
+        self.customer.trace_add("write", self.make_name)
+        self.client.trace_add("write", self.make_name)
+        self.field.trace_add("write", self.make_name)
+        self.sample.trace_add("write", self.make_name)
 
         # set defaults
         # todo #3 abstract these out into some TOML or something
@@ -83,10 +81,12 @@ class Project:
             name = f"{name} ({self.sample.get()})".strip()
         self.name.set(name)
 
-    @staticmethod
-    def dump_json(project: Project) -> None:
-        """Dump a JSON representation of a Project at the passed path."""
-        for test in project.tests:
+    def dump_json(self, path=None) -> None:
+        """Dump a JSON representation of the Project at the passed path."""
+        if path is None:
+            path = self.path.get()
+
+        for test in self.tests:
             if test.chemical.get().strip() != test.chemical.get():
                 test.chemical.set(test.chemical.get().strip())
 
@@ -97,7 +97,7 @@ class Project:
         _blanks = {}  # put in dicts to allow for popping later
         _trials = {}
         keys = []
-        for test in project.tests:
+        for test in self.tests:
             key = test.label.get().lower()  # eliminate capitalization discrepancies
             while key in keys:  # checking for duplicate values
                 test.label.set(test.label.get() + " - copy")
@@ -121,47 +121,46 @@ class Project:
         )
         trials = [_trials.pop(name) for name in trial_labels]
 
-        project.tests = [*blanks, *trials]
+        self.tests = [*blanks, *trials]
 
         this = {
             "info": {
-                "customer": project.customer.get(),
-                "submittedBy": project.submitted_by.get(),
-                "productionCo": project.client.get(),
-                "field": project.field.get(),
-                "sample": project.sample.get(),
-                "sampleDate": project.sample_date.get(),
-                "recDate": project.received_date.get(),
-                "compDate": project.completed_date.get(),
-                "name": project.name.get(),
-                "analyst": project.analyst.get(),
-                "numbers": project.numbers.get(),
-                "path": os.path.abspath(project.path.get()),
-                "notes": project.notes.get(),
+                "customer": self.customer.get(),
+                "submittedBy": self.submitted_by.get(),
+                "productionCo": self.client.get(),
+                "field": self.field.get(),
+                "sample": self.sample.get(),
+                "sampleDate": self.sample_date.get(),
+                "recDate": self.received_date.get(),
+                "compDate": self.completed_date.get(),
+                "name": self.name.get(),
+                "analyst": self.analyst.get(),
+                "numbers": self.numbers.get(),
+                "path": os.path.abspath(self.path.get()),
+                "notes": self.notes.get(),
             },
             "params": {
-                "bicarbonates": project.bicarbs.get(),
-                "bicarbsIncreased": project.bicarbs_increased.get(),
-                "chlorides": project.chlorides.get(),
-                "baseline": project.baseline.get(),
-                "temperature": project.temperature.get(),
-                "limitPSI": project.limit_psi.get(),
-                "limitMin": project.limit_minutes.get(),
-                "interval": project.interval.get(),
-                "uptake": project.uptake.get(),
+                "bicarbonates": self.bicarbs.get(),
+                "bicarbsIncreased": self.bicarbs_increased.get(),
+                "chlorides": self.chlorides.get(),
+                "baseline": self.baseline.get(),
+                "temperature": self.temperature.get(),
+                "limitPSI": self.limit_psi.get(),
+                "limitMin": self.limit_minutes.get(),
+                "interval": self.interval.get(),
+                "uptake": self.uptake.get(),
             },
-            "tests": [test.dump_json() for test in project.tests],
-            "outputFormat": project.output_format.get(),
-            "plot": os.path.abspath(project.plot.get()),
+            "tests": [test.dump_json() for test in self.tests],
+            "outputFormat": self.output_format.get(),
+            "plot": os.path.abspath(self.plot.get()),
         }
 
-        with open(project.path.get(), "w") as file:
+        with open(path, "w") as file:
             json.dump(this, file, indent=4)
 
-        logger.info("Saved %s to %s", project.name.get(), project.path.get())
+        logger.info("Saved %s to %s", self.name.get(), path)
 
-    @staticmethod
-    def load_json(path: str) -> Project:
+    def load_json(self, path: str) -> None:
         """Return a Project from a passed path to a JSON dump."""
         logger.info("Loading from %s", path)
         with open(path, "r") as file:
@@ -174,39 +173,36 @@ class Project:
             )
             obj["info"]["path"] = path
 
-        this = Project()
         info = obj.get("info")
-        this.customer.set(info.get("customer"))
-        this.submitted_by.set(info.get("submittedBy"))
-        this.client.set(info.get("productionCo"))
-        this.field.set(info.get("field"))
-        this.sample.set(info.get("sample"))
-        this.sample_date.set(info.get("sampleDate"))
-        this.received_date.set(info.get("recDate"))
-        this.completed_date.set(info.get("compDate"))
-        this.name.set(info.get("name"))
-        this.numbers.set(info.get("numbers"))
-        this.analyst.set(info.get("analyst"))
-        this.path.set(info.get("path"))
-        this.notes.set(info.get("notes"))
+        self.customer.set(info.get("customer"))
+        self.submitted_by.set(info.get("submittedBy"))
+        self.client.set(info.get("productionCo"))
+        self.field.set(info.get("field"))
+        self.sample.set(info.get("sample"))
+        self.sample_date.set(info.get("sampleDate"))
+        self.received_date.set(info.get("recDate"))
+        self.completed_date.set(info.get("compDate"))
+        self.name.set(info.get("name"))
+        self.numbers.set(info.get("numbers"))
+        self.analyst.set(info.get("analyst"))
+        self.path.set(info.get("path"))
+        self.notes.set(info.get("notes"))
 
         params = obj.get("params")
-        this.bicarbs.set(params.get("bicarbonates"))
-        this.bicarbs_increased.set(params.get("bicarbsIncreased"))
-        this.chlorides.set(params.get("chlorides"))
-        this.baseline.set(params.get("baseline"))
-        this.temperature.set(params.get("temperature"))
-        this.limit_psi.set(params.get("limitPSI"))
-        this.limit_minutes.set(params.get("limitMin"))
-        this.interval.set(params.get("interval"))
-        this.uptake.set(params.get("uptake"))
+        self.bicarbs.set(params.get("bicarbonates"))
+        self.bicarbs_increased.set(params.get("bicarbsIncreased"))
+        self.chlorides.set(params.get("chlorides"))
+        self.baseline.set(params.get("baseline"))
+        self.temperature.set(params.get("temperature"))
+        self.limit_psi.set(params.get("limitPSI"))
+        self.limit_minutes.set(params.get("limitMin"))
+        self.interval.set(params.get("interval"))
+        self.uptake.set(params.get("uptake"))
 
-        this.plot.set(obj.get("plot"))
-        this.output_format.set(obj.get("outputFormat"))
+        self.plot.set(obj.get("plot"))
+        self.output_format.set(obj.get("outputFormat"))
 
         for entry in obj.get("tests"):
             test = Test()
             test.load_json(entry)
-            this.tests.append(test)
-
-        return this
+            self.tests.append(test)