+++ CHANGELOG.md
@@ -13,13 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - rinse windows will be titled after the system they control
  - requests for rinse cycles will be ignored if a test is already running
  - the Live Plot stops updating (clearing itself) at the end of a test
+ - evaluation windows will title themselves after their test handler and project
+ - minor performance improvement to evaluation window
 ### Fixed
  - observed baseline pressure reporting
 ### Removed
  - dependency on openpyxl
  - dependency on PIL
  - ability to export report as .xlsx
- - redundant windows from EvaluationFrame
+ - redundant tabs from EvaluationFrame
 
 ## [v0.1.0]
 ### Added